backend/requirements.txt
Django==2.0.5
djangorestframework==3.8.2
django-cors-headers==2.2.0
